Load initrd pointer from dtb

+++ lab2/include/uart.h
@@ -9,6 +9,6 @@ void writeline_uart(const char *buffer, int count);
 int interact_readline_uart(char *buffer);
 void puts_uart(const char *buffer);
 void print_uart(const char *buffer);
-void write_num_uart(unsigned num);
+void write_num_uart(unsigned long num);
 
 #endif
